refactor: extract `argus-dispatch` crate
