+++ Cargo.toml
@@ -8,6 +8,7 @@ resolver = "2"
 [workspace.dependencies]
 alloy = { version = "1.0.37", features = ["essentials"] }
 argus-core = { path = "crates/argus-core" }
+argus-dispatch = { path = "crates/argus-dispatch" }
 argus-providers = { path = "crates/argus-providers" }
 argus-store = { path = "crates/argus-store" }
 omnihook = { path = "crates/omnihook" }
@@ -69,6 +70,7 @@ edition = "2024"
 [dependencies]
 alloy.workspace = true
 argus-core.workspace = true
+argus-dispatch.workspace = true
 argus-providers.workspace = true
 argus-store.workspace = true
 arc-swap.workspace = true
@@ -114,6 +116,7 @@ urlencoding.workspace = true
 
 [dev-dependencies]
 argus-core = { path = "crates/argus-core", features = ["test-utils"] }
+argus-dispatch.workspace = true
 argus-store = { path = "crates/argus-store" }
 mockall.workspace = true
 mockito.workspace = true
