+++ crates/argus-dispatch/src/publisher/error.rs
