+++ crates/argus-dispatch/src/publisher/traits.rs
