+++ crates/argus-dispatch/src/template/mod.rs
