+++ src/context/mod.rs
@@ -24,16 +24,15 @@ pub use metrics::{AppMetrics, Metrics};
 
 use crate::{
     abi::{AbiService, repository::AbiRepository},
-    action_dispatcher::{ActionDispatcher, template::TemplateService},
     engine::{
         alert_manager::AlertManager,
         filtering::RhaiFilteringEngine,
         rhai::{RhaiCompiler, RhaiScriptValidator},
     },
-    http_client::HttpClientPool,
     loader::load_config,
     monitor::{MonitorManager, MonitorValidator},
 };
+use argus_dispatch::{ActionDispatcher, http_client::HttpClientPool, template::TemplateService};
 use argus_providers::rpc::create_provider;
 use argus_store::SqliteStateRepository;
