+++ src/engine/outbox_processor.rs
@@ -10,7 +10,7 @@ use argus_core::{
 use futures::{StreamExt, stream};
 use tokio_util::sync::CancellationToken;
 
-use crate::action_dispatcher::ActionDispatcher;
+use argus_dispatch::ActionDispatcher;
 
 /// Service responsible for processing the persistent action queue.
 pub struct OutboxProcessor<S: AppRepository + ?Sized> {
