+++ src/lib.rs
@@ -6,11 +6,9 @@
 
 pub mod abi;
 pub mod action;
-pub mod action_dispatcher;
 pub mod cmd;
 pub mod context;
 pub mod engine;
-pub mod http_client;
 pub mod http_server;
 pub mod loader;
 pub mod monitor;
