+++ src/test_helpers/http_client.rs
@@ -4,7 +4,7 @@ use argus_core::config::HttpRetryConfig;
 use reqwest::Client;
 use reqwest_middleware::ClientWithMiddleware;
 
-use crate::http_client::create_retryable_http_client;
+use argus_dispatch::http_client::create_retryable_http_client;
 
 /// Creates a default HTTP client with retry capabilities for testing purposes.
 pub fn create_test_http_client() -> Arc<ClientWithMiddleware> {
